fix: 2026-06-22/27 버그수정·기능추가 일괄 적용
- [반입실적] 반입량 단위 ton → kg 표시 변환 (IntakeTab 전체)
- [연월차] INTERNAL_ADMIN 휴가결재 서명 미저장 수정 (useStoredSignature 조건 보정)
- [근태] 퇴근 초기화 기능 추가 (isAttendanceManager 동일 권한, checkOutTime 존재 시 표시)
- [사용자관리] 집하시설 없는 지자체 위탁업체 등록 제한 (no_facilities 400 반환)
- [민원] 관리자 등록 민원 사진 근로자 화면 미표시 수정 (requestImage/completionImage 분리)
- [휴가결재] INTERNAL_ADMIN not_found 수정 (userScope findFirst 분리 패턴 적용)
- [날씨] 날씨관리대장 출력·사진 업로드 기능 개선
- [차량일지] PDF 생성 및 제출 API 개선
- prisma 마이그레이션 2건 추가 (vehicle_pdf_archive, weather_notice_photo)

+++ app/api/safety/weather-notices/[id]/photos/route.ts
@@ -17,7 +17,7 @@ function isManager(role: string) {
 const PostBody = z.object({
   photoData:   z.string().max(400_000, '사진 크기가 너무 큽니다. 더 작은 사진을 사용해 주세요.').optional().nullable(),
   recordTime:  z.string().regex(/^\d{2}:\d{2}$/).optional().nullable(),
-  feelsLike:   z.number().int().min(-50).max(60).optional().nullable(),
+  feelsLike:   z.number().min(-50).max(60).optional().nullable(),
   actionTaken: z.string().trim().max(1000).optional().nullable(),
   managerName: z.string().trim().max(50).optional().nullable(),
 });
